% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{plot.questionlink_prepare}
\alias{plot.questionlink_prepare}
\title{Plot function plotting found connections between questions}
\usage{
\method{plot}{questionlink_prepare}(
  object,
  major_breaks_divisor = 10,
  minor_breaks_divisor = 1,
  combine_population_plots = T
)
}
\description{
Plot function plotting found connections between questions
}
