% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{print.questionlink_harmonize}
\alias{print.questionlink_harmonize}
\title{Print information on ql_harmonize object}
\usage{
\method{print}{questionlink_harmonize}(x, ...)
}
\arguments{
\item{x}{An object of type "questionlink_harmonize"}
}
\description{
Displays progress in the questionlink workflow and a highlevel overview
over the number of found connections.
}
