% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{print.questionlink_prepare}
\alias{print.questionlink_prepare}
\title{Print information on ql_prepare object}
\usage{
\method{print}{questionlink_prepare}(x, ...)
}
\arguments{
\item{x}{An object of type "questionlink_prepare"}

\item{...}{Not used.}
}
\description{
Displays progress in the questionlink workflow and a highlevel overview
over the number of found connections.
}
