% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ql_check.R
\name{print_has_no_negative}
\alias{print_has_no_negative}
\title{Prints the warning whether negative values were found.}
\usage{
print_has_no_negative(question, question_information, has_no_negative_flag)
}
\description{
Prints the warning whether negative values were found.
}
