% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ql_check.R
\name{print_is_continuous}
\alias{print_is_continuous}
\title{Prints the warning whether all response options between min and max were found}
\usage{
print_is_continuous(question, question_information, is_continuous_flag)
}
\description{
Prints the warning whether all response options between min and max were found
}
