% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ql_check.R
\name{print_is_singular}
\alias{print_is_singular}
\title{Prints the warning whether more than one version of the question is found}
\usage{
print_is_singular(question, question_information, is_singular_flag)
}
\description{
Prints the warning whether more than one version of the question is found
}
