% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ql_check.R
\name{ql_check}
\alias{ql_check}
\title{Checks the supplied data for common pitfalls in the harmonization process}
\usage{
ql_check(ql_prepare_object)
}
\description{
ql_check() performs plausbility checks on the supplied data. It prints messages
if inconsistencies in the data are found which may influence the results of the harmonization.
In particular, three checks are performed:
}
\section{Check if multiple versions of the same question are found}{

Each question should be associated with one set of response options. If a
one set of response options was found for one year (e.g. 1,2,3,4 in the year 2000) and another set was
found in another year (e.g. 1,2,3 in  the year 2004) there may be a problem with the data, such as missing cases
or falsly exculded cases. Make sure to include only one version of the question, as the harmonization may lead to wrong
results.
}

\section{Check if negative responses are found}{

Negative values are used to represent missing values of some sort in most survey data files.
Check whether negative values are valid responses in your datafile (and ignore the message if they are valid).
}

\section{Check if response options are not used in a scale}{

Missing response options (e.g. in a 4 point scale 1,2,4, response option 3 would be missing), can lead to
false harmonization results. If response options are missing, check if responses were falsely excluded from the
dataset. This check utilizes the information supplied by the 'scale_min_max' argument in ql_prepare().
If scale_min_max is left empty, the minimum and maximum found in the data will be used.
}

