% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ql_harmonize.R
\name{ql_harmonize}
\alias{ql_harmonize}
\title{Harmonize ql_prepare object}
\usage{
ql_harmonize(ql_prepare_object)
}
\arguments{
\item{ql_prepare_object}{An object of type questionlink_prepare, created by 'ql_prepare()'}
}
\value{
An object of type 'questionlink' 'questionlink_harmonize'
}
\description{
Harmonizes all harmonization opportunities that were found with 'ql_prepare()' by creating recoding
tables for each harmonization opportunity.
}
\details{
This function is step 2 of 3 in the questionlink harmonization workflow.

Depending on the number of harmonization opportunities, this may take some time.
}
\examples{
ql_prepare(data) |> ql_harmonize()
}
