% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ql_prepare.R
\name{ql_prepare}
\alias{ql_prepare}
\title{Prepare data for harmonization}
\usage{
ql_prepare(
  data,
  use_relay = F,
  allowed_direct_connections = NULL,
  allowed_relay_connections = NULL,
  time_relaxation = 0,
  remove_circular_relays = TRUE,
  scale_min_max = list()
)
}
\arguments{
\item{data}{A tibble/data.frame containing the data in a format outlined in the details below.}

\item{use_relay}{A logical indicating whether relay connections should be calculated. Default = FALSE.}

\item{allowed_direct_connections}{Further filtering of direct connections. See "Finding opportunities to harmonize" in the details.}

\item{allowed_relay_connections}{Further filtering of relay connections. See "Finding opportunities to harmonize" in the details.}

\item{time_relaxation}{A numeric which specifies how much time is allowed to differ to establish direct connections. Default = 0.}

\item{remove_circular_relays}{Logical whether circular relays (relays from the same source question to the same target question
should be removed). Default is TRUE.}

\item{scale_min_max}{A named list which contains numeric vectors of length two, e.g. 'list("Question A" = c(0,5))'.
The name of the numeric vector should correspond to one
of the questions supplied in 'data'.The first element of
the numeric vector is the minimum of a scale,
the second element is the maximum of a scale.
Should be supplied if not all possible response options were used
by respondents. See Details for more information.}
}
\value{
An object of type 'questionlink' 'questionlink_prepare'.
}
\description{
The function prepares the data for harmonization with Observed Score Equating
in a Random Groups design (OSE-RG) by finding all possible opportunities for harmonization in the data.
}
\details{
OSE-RG can be used when data from two question measuring the same construct are
used in the same population at roughly similar times (e.g. the same year).

The data has to be supplied as a tibble/data.frame in a "long format".
The data.frame/tibble has to have the following columns names and variable types:
\itemize{
\item year: Numeric variable which specifies the year in which the survey was conducted. Can not be of type Date or Factor. If other units of time are needed look here.
\item question: Character variable which specifies which kind of question was asked.
\item response: Numeric variable which contains the response values. NAs and missings should be removed.
Note that the responses should be recoded so that higher values represent the same meaning on the scale.
\item (optional) weight: Numeric variable which contains the weight associated with a response.
\item (optional) population: Character variable which describes the population a response is attributed to.
Data should only be harmonized within one population. If no population is supplied
all data will be treated to originate from the same population.
}
}
\section{Finding opportunities to harmonize}{

}

\examples{
ql_prepare(data = x,
           use_relay = T,
           allowed_direct_connections = list(year_source == year_target,
                                             instrument_id_source == instrument_id_target

           allowed_relay_connections = list(instrument_id_relay_source != instrument_id_relay_target,
                                             insturment_id_source != instrument_id_relay_target))
}
