% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{questionlink_example_data}
\alias{questionlink_example_data}
\title{Questionlink example dataset}
\format{
\subsection{\code{questionlink_example_data}}{

A data frame with 328,00 rows and 4 columns:
\describe{
\item{year}{A year in numeric format}
\item{question}{A question identifier.}
\item{response}{Scale responses from respondents (in this case simulated).}
\item{weight}{A column with respondent specific weights. In this case all 1.}
}
}
}
\usage{
questionlink_example_data
}
\description{
An example dataset which is in the correct format for questionlink functions.
It is also used in the vignette "questionlink_tutorial".
}
\keyword{datasets}
