% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{summary.questionlink_harmonize}
\alias{summary.questionlink_harmonize}
\title{Summary function, printing full list connections}
\usage{
\method{summary}{questionlink_harmonize}(object, ...)
}
\description{
Summary function, printing full list connections
}
