% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{summary.questionlink_prepare}
\alias{summary.questionlink_prepare}
\title{Summarizes question and connection information found in the data.}
\usage{
\method{summary}{questionlink_prepare}(object, ...)
}
\arguments{
\item{object}{A questionlink_prepare object.}

\item{...}{Not used.}
}
\description{
Summarizes question and connection information found in the data.
}
