% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ql_transform.R
\name{ql_transform}
\alias{ql_transform}
\title{Transform data using questionlink recoding tables}
\usage{
ql_transform(ql_harmonize_object, target_question, new_data)
}
\arguments{
\item{ql_harmonize_object}{An object of type questionlink_harmonize, created by 'ql_harmonize()'}

\item{target_question}{A character value specifying the question which serves as the transformation target.}

\item{new_data}{A data frame or tibble which contains the following columns "question", "response". The data frame
also needs to contain the column "population", if the column was also used in the 'ql_prepare()' step.}
}
\value{
An object of type 'questionlink' 'questionlink_harmonize'
}
\description{
Transforms data using recoding tables obtained by using 'ql_prepare() |> ql_harmoninze().
}
\details{
This function is step 3 of 3 in the questionlink harmonization workflow.
This function either transforms data which is supplied in the harmonization workflow, or transformes
new data, which can be added via the "new_data" argument.

Circular relays:

If circular relays are set to TRUE (default is FALSE) during the invocation of ql_prepare(), the results of ql_transform()
contain the numerical equivalents that were found using the relays. If no circular relays are used, only the
results of ql_transformed contain the same response options for the target questions as were supplied in the data.
}
\examples{
ql_prepare(data) |> ql_harmonize() |> ql_transform()
}
