% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{summary.questionlink_prepare}
\alias{summary.questionlink_prepare}
\title{Summary function, printing full list connections}
\usage{
\method{summary}{questionlink_prepare}(object, ...)
}
\description{
Summary function, printing full list connections
}
